requirements.txt
# Product Search Dependencies
requests>=2.28.0
beautifulsoup4>=4.11.0
python-dotenv>=1.0.0
PyYAML>=6.0
